API requests is divided into separate files
