+++ server/dev/DevMode
@@ -1,0 +1,8 @@
+const express = require('express')
+const app = express
+const cors = require('cors')
+
+
+if (process.env.NODE_ENV === 'development'){
+    app.use(cors())
+    }
