+++ server/group/Settings of Server/static.js
@@ -1,0 +1,3 @@
+function dist (app, express) {
+    app.use('/', express.static('dist'));
+} module.exports.dist = dist
