+++ server/group/Settings of Server/useCors.js
@@ -1,0 +1,8 @@
+const cors = require('cors')
+
+function useCors(app) {
+    if (process.env.NODE_ENV === 'development'){
+    app.use(cors())
+    }
+}
+module.exports.useCors = useCors
