+++ server/package.json
@@ -17,6 +17,7 @@
     "express": "^4.17.3",
     "express-session": "^1.17.2",
     "font-awesome": "^4.7.0",
+    "morgan": "^1.10.0",
     "pg": "^8.7.3",
     "prod": "^1.0.1",
     "run": "^1.4.0"
